See Description for all the details I've learned for this chapter
property binding:

if the property would be enclosed with square brackets, you dont need to use the expression to evaluate the property value 
eg: 
<img [src] = "person.imgPath"/>

thus, if you are going to assign a value to a property without a square brackets, you need to use the expression to evaluate the property value

<img src = "{{person.imgPath}}"/>


attribute binding:
if you are going to bind a property to an element's attribute you need to have attr. as a prefix to the attribute you're going to assign with a value

<td  [attr.colspan]="table.colspan"></td>

class Binding:
if you want to add a class dynamically with respect with your indicators. , this must be possible with style binding, you just channge the class binding syntax to .. [style.stylename]='indicator'

you can add class by.. having [class.CLASSNAME]="propertyName"

eg. [class.active] = '!person.isResigned'


how to add bootstrap in your angular project?

-install bootstrap with npm
->npm install bootstrap --save
--> then go to styles.css
    @import(PATH)



event Binding.
one of the most popular event you're going to use is (click).

for you to bind an event to click. define a function in the component.ts of your desired component

function btnClick($event){
   $event.stopPropagation();
  console.log('btn clicked');
}
so you're wondering why is there a need for the line of $event.stopPropagation() to be present in the function..

for example you have a nested elements, and both of them have an event, if you trigger the inner element's event, the event will bubble up with the outer element. that means multiple commands and isnt necessary
div>btn
<div>
   <button></button>
</div>




so, in other inputs you can have a enter trigger or a specific character. you can access it after the keyup event..

(keyup.enter)= 'method'


#TemplateVariable.. if you want to have a variable inline with your html element, it is possible
<input #email (keyup.enter)='sendEmail(email)'/>


##you cant use TemplateVariable as always, sometimes you need to have two-way binding for your forms. and inputs. you just need to remember the banana on the box [(ngModel)]="propertyName"

also two-way binding or [(ngModel)] wont work because FORMS isnt yet present in the app.module.ts

import {FormsModule} from "@angular/forms";
and add FormsModule in the imports object