see the description for the explanation of pipe
Where the Pipe can be used?


- it is used to display data with corresponding arrangement, or signature

here are the common pipes that is available in angular
1.uppercase
2. lowercase
3. number
4. date
5. currency
6. json, etc


how to manipulate the display of number?
{{course.price | currency:'CurrenyCode':[isDisplayed]:'[number of wholenumber].[number of decimal places]'}}

working Example:
{{course.price | currency:'PHP':true:'1.2-3'}}

explanation, the pipe is currency, and the currencyCode is PHP and it must be displayed, it should have atlease 1 whole number, and 2-3 decimal places if decimal places exceed, it is rounded to the end of the range


for the other formats of date, you can check at the document of angular
{{course.releaseDate| date:'[format]'}}
      {{course.releaseDate|date:'shortDate'}}